pyproject.toml
[project]
name = "hallucination-creation"
version = "0.1.0"
description = "Scripted pipeline for generating hallucination detection datasets."
authors = [{ name = "Hallucination Creation Team" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic-ai>=0.0.12",
    "openai>=1.44.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
data = [
    "pandas>=2.2.0",
    "pyarrow>=16.1.0",
]
test = [
    "pytest>=8.2.0",
]

[tool.uv]
dev-dependencies = ["pytest>=8.2.0"]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
